Refactor CreditChecker; fix check for user credits
